+++ backend/controllers/protocolController.js
@@ -1,9 +1,11 @@
 const asyncHandler = require('express-async-handler');
 const Protocol = require('../models/protocolModel');
+const User = require('../models/userModel');
 
 // @desc Get protocols
 // @route GET /api/protocols
-// @access private
+// @access public
+// Ok to return user?
 const getProtocols = asyncHandler(async (req, res) => {
     const protocols = await Protocol.find();
     res.status(200).json(protocols);
@@ -11,7 +13,8 @@ const getProtocols = asyncHandler(async (req, res) => {
 
 // @desc Get protocol by id
 // @route GET /api/protocols/:id
-// @access private
+// @access public
+// Is this needed?
 const getProtocolById = asyncHandler(async (req, res) => {
     const protocol = await Protocol.findById(req.params.id);
 
@@ -20,7 +23,7 @@ const getProtocolById = asyncHandler(async (req, res) => {
         throw new Error("Protocol not found");
     }
 
-    res.status(200).json(protocol);
+    res.status(200).json(protocol.select('-user'));
 });
 
 // @desc Set protocol
@@ -31,8 +34,10 @@ const setProtocol = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error('Please enter a name field');
     }
+
     const protocol = await Protocol.create({
-        name: req.body.name
+        name: req.body.name,
+        user: req.user.id
     });
 
     res.status(200).json(protocol);
@@ -47,6 +52,18 @@ const updateProtocol = asyncHandler(async (req, res) => {
     if (!protocol) {
         res.status(400);
         throw new Error ('Protocol not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+
+    if (protocol.user.toString() !== user.id){
+        res.status(401);
+        throw new Error("User not authorized");
     }
 
     const updatedProtocol = await Protocol.findByIdAndUpdate(req.params.id, req.body, {
@@ -66,6 +83,18 @@ const deleteProtocol = asyncHandler(async (req, res) => {
         res.status(400);
         throw new Error ('Protocol not found');
     };
+
+    const user = await User.findById(req.user.id);
+
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+
+    if (protocol.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
+    }
 
     await protocol.deleteOne();
